requirements.txt
pyserial==3.5
numpy>=1.20.0
tqdm>=4.60.0 